fix: utilisation de la class Point
